Refactor EntityId component and add generateEntityId function
